requirements.txt
transitions
asyncpg
pydantic
python-dotenv
pydantic_settings